move files to src
