+++ web/src/pages/_app.tsx
@@ -2,9 +2,9 @@ import Head from 'next/head';
 import { ThemeProvider } from '@mui/material/styles';
 import CssBaseline from '@mui/material/CssBaseline';
 import { CacheProvider } from '@emotion/react';
-import createEmotionCache from '../styles/createEmotionCache';
-import "../styles/globals.css";
-import theme from '../styles/theme';
+import createEmotionCache from 'styles/createEmotionCache';
+import "styles/globals.css";
+import theme from 'styles/theme';
 import type { AppProps } from "next/app";
 import { ChakraProvider } from "@chakra-ui/react";
 import { type EmotionCache } from '@emotion/cache/types';
