+++ web/src/pages/_document.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import Document, { Html, Head, Main, NextScript, DocumentContext } from 'next/document';
 import createEmotionServer from '@emotion/server/create-instance';
-import createEmotionCache from '../styles/createEmotionCache';
+import createEmotionCache from 'styles/createEmotionCache';
 
 export default class MyDocument extends Document {
   render() {
